# src/main/resources/openapi.yml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Subito Purchase Cart Service API",
    "description": "RESTful API for managing orders with a shopping basket system",
    "contact": {
      "name": "Subito Development Team",
      "email": "dev@subito.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/v1/orders": {
      "post": {
        "tags": [
          "order-controller"
        ],
        "operationId": "createOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/orders/{orderId}/ship": {
      "post": {
        "tags": [
          "order-controller"
        ],
        "operationId": "shipOrder",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShipOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/orders/{orderId}": {
      "get": {
        "tags": [
          "order-controller"
        ],
        "operationId": "getOrder",
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateOrderRequest": {
        "type": "object",
        "properties": {
          "basketId": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "lastName": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 1
          },
          "phoneNumber": {
            "type": "string",
            "minLength": 1,
            "pattern": "^\\+?[0-9]+$"
          },
          "street": {
            "type": "string",
            "minLength": 1
          },
          "city": {
            "type": "string",
            "minLength": 1
          },
          "postalCode": {
            "type": "string",
            "minLength": 1,
            "pattern": "^[0-9]+$"
          },
          "country": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "basketId",
          "city",
          "country",
          "email",
          "lastName",
          "name",
          "phoneNumber",
          "postalCode",
          "street"
        ]
      },
      "OrderItemResponse": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int64"
          },
          "productName": {
            "type": "string"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "unitNetPrice": {
            "type": "number"
          },
          "price": {
            "type": "number"
          },
          "vat": {
            "type": "number"
          },
          "total": {
            "type": "number"
          }
        }
      },
      "OrderResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string",
            "enum": [
              "INSERTED",
              "PAID",
              "SHIPPED",
              "EXPIRED"
            ]
          },
          "insertDate": {
            "type": "string",
            "format": "date-time"
          },
          "totalPrice": {
            "type": "number"
          },
          "totalVat": {
            "type": "number"
          },
          "grandTotal": {
            "type": "number"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItemResponse"
            }
          }
        }
      },
      "ShipOrderRequest": {
        "type": "object",
        "properties": {
          "trackingUrl": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "trackingUrl"
        ]
      },
...
    }
  }
}